requirements.txt
numpy
opencv-python
mediapipe
pyvirtualcam
pyyaml
toml
loguru